Added Windows button event, polished event queue.
Win32 will need a bit more TLC, there's a few parts of the API that need clarification such as input to string, should we give the name of the input, or the result of it, so "Enter" vs backslash n. The name makes the most sense, but where should we handle the other case?